# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF645 - Service Qualification\n\n### Release\
    \ : 18.5 - December 2018\n\nService Qualification API is one of Pre-Ordering Management\
    \ API Family. Service Qualification API goal is to provide service technical eligibility.\
    \ \n\n### Resource \n- serviceQualification\n\n### Operations\nService Qualification\
    \ API performs the following operations on the resource :\n- Retrieve an entity\
    \ or a collection of entities depending on filter criteria\n- Partial update of\
    \ an entity (including updating rules)\n- Create an entity (including default\
    \ values and creation rules)\n- Delete an entity (for administration purposes)\n\
    - Manage notification of events"
  version: "4.0.0"
  title: "Service Qualification Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /serviceQualification:
    get:
      tags:
      - "serviceQualification"
      summary: "List or find ServiceQualification objects"
      description: "This operation list or find ServiceQualification entities"
      operationId: "listServiceQualification"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceQualification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceQualification"
    post:
      tags:
      - "serviceQualification"
      summary: "Creates a ServiceQualification"
      description: "This operation creates a ServiceQualification entity."
      operationId: "createServiceQualification"
      parameters:
      - in: "body"
        name: "serviceQualification"
        description: "The ServiceQualification to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceQualification_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceQualification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceQualification"
  /serviceQualification/{id}:
    get:
      tags:
      - "serviceQualification"
      summary: "Retrieves a ServiceQualification by ID"
      description: "This operation retrieves a ServiceQualification entity. Attribute\
        \ selection is enabled for all first level attributes."
      operationId: "retrieveServiceQualification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceQualification"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceQualification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceQualification"
    delete:
      tags:
      - "serviceQualification"
      summary: "Deletes a ServiceQualification"
      description: "This operation deletes a ServiceQualification entity."
      operationId: "deleteServiceQualification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceQualification"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceQualification"
    patch:
      tags:
      - "serviceQualification"
      summary: "Updates partially a ServiceQualification"
      description: "This operation updates partially a ServiceQualification entity."
      operationId: "patchServiceQualification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceQualification"
        required: true
        type: "string"
      - in: "body"
        name: "serviceQualification"
        description: "The ServiceQualification to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceQualification_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceQualification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceQualification"
...
definitions:
...
  RelatedParty:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of a related party"
      href:
        type: "string"
        description: "Reference of the related party, could be a party reference or\
          \ a party role reference"
      name:
        type: "string"
        description: "Name of the related party"
      role:
        type: "string"
        description: "Role of the related party."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "RelatedParty reference. A related party defines party or party role\
      \ linked to a specific entity."
    example:
      '@referredType': "@referredType"
      role: "role"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
...
  ServiceQualification_Update:
    type: "object"
    properties:
      description:
        type: "string"
        description: "Description of the serviceQualification"
      estimatedResponseDate:
        type: "string"
        format: "date-time"
        description: "Date when the requester expect to provide an answer for the\
          \ qualification request."
      expectedQualificationDate:
        type: "string"
        format: "date-time"
        description: "A date (DateTime). Deadline date when the requester expected\
          \ a qualification answer."
      expirationDate:
        type: "string"
        format: "date-time"
        description: "Date when the qualification response expires"
      externalId:
        type: "string"
        description: "Identifier provided by the requester"
      provideAlternative:
        type: "boolean"
        description: "When the value is TRUE means that alternative solutions should\
          \ be provided"
      provideOnlyAvailable:
        type: "boolean"
        description: "When the value is TRUE means that only available service must\
          \ be listed in the response"
      provideUnavailabilityReason:
        type: "boolean"
        description: "When the value is TRUE means that unavailability reason are\
          \ expected for non available service."
      qualificationResult:
        type: "string"
        description: "Qualification result for this serviceQualification. It could\
          \ be:  qualified (all qualification item are qualified), alternate (At least\
          \ one item alternate and no item with  unqualified), unqualified (At least\
          \ one item unqualified)"
      state:
        type: "string"
        description: "State of the serviceQualification (acknowledged, inProgress,\
          \ terminatedWithError, done)"
      relatedParty:
        type: "array"
        description: "A list of related party references, defines party or party role\
          \ linked to this request."
        items:
          $ref: "#/definitions/RelatedParty"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "ServiceQualification is used to perform a technical eligibility.\
      \ It allows to retrieve a list of services that are technically available in\
      \ the context of the interaction (place, party, service characteristics, ...).\n\
      Skipped properties: id,href,effectiveQualificationDate,serviceQualificationDate,serviceQualificationItem"
    example:
      expectedQualificationDate: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      description: "description"
      externalId: "externalId"
      provideAlternative: true
      provideUnavailabilityReason: true
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      provideOnlyAvailable: true
      qualificationResult: "qualificationResult"
      '@baseType': "@baseType"
      state: "state"
      '@schemaLocation': "http://example.com/aeiou"
      estimatedResponseDate: "2000-01-23T04:56:07.000+00:00"
      expirationDate: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"